+++ Assets/VRM10/Editor/Components/VRM10ControllerSceneView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc70b88276e055640aa0da2d937c8f39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
